Modify restart logic to actually restart
The previous version of this code didn't work because
the restart command was being run in a subshell of the
process we were trying to restart. Thin ended up stopping
that process but never starting a new one.

By switching to fork and detaching the child we let thin
stop the original process and restart a new one.