Persist proxy config to file (STEAM-94)
The configure command no longer forces AS to start when
configuring the proxy list. Instead, it writes the new proxies
to /etc/sysconfig/jboss-as and signals a restart of AS. If
AS was already running, it still writes out the file but also
twiddles the new values to the running AS w/o requiring a restart.