Namespace common command names in module [STEAM-129]
The BaseService now expects to see the deploy, undeploy, and
configure commands all under a module named after the service.

So, SteamCannon::JBossASService commands should be under
SteamCannon::JBossAS::{Configure,Deploy,Undeploy}Service.

It probably makes sense to move *all* commands in to modules
for consistency, but as a minimum we need to move common named
commands at least to prevent namespace collisions.