Postgresql service now configures the EBS volume and inits postgresql [STEAM-128]
The service spins and waits for the EBS device to appear,
then tries to mount it. If it is mountable, that means it has 
already been formatted, so we don't format again. Otherwise,
the volume gets formatted as ext3. 

The service then checks to see if there is a pg_hba.conf file exists
on the volume. If it is not there, it initializes postgresql
(via service postgresql initdb).