Agent support for securing AS jmx and admin consoles (STEAM-84)
Similar to the postgresql service, the jboss_as service now accepts
{ "create_admin": {"user":"user", "password":"password"}}
in its configure body to set the admin username and password for
/jmx-console and /admin-console web apps.

This change also fixes some broken logic in the jboss_as configure
command. Previously, it would only restart AS if the last sub-command
called wanted to restart. Now it will restart if any of those sub-commands
need a restart.